Do not change job parameters with env variables
Had a problem that every time when running a job
with environment variables for execPattern,
classPattern and sourcePattern the job configuration
would change to the expanded values. Which isn't that
good when the build is parameterized and can change
from build to build.

Also doing the expansion a little bit earlier so that
log messages contains the expanded values.

Added testcase
JacocoPublisherTest.testPerformWithEnvVarsInSettings
that would fail without the changes in
JacocoPublisher.java